requirements.txt
Django>=5.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0 